make reducer
