+++ package.json
@@ -15,6 +15,7 @@
     "react-scripts": "3.0.1",
     "redux": "^4.0.1",
     "redux-devtools-extension": "^2.13.8",
+    "redux-thunk": "^2.3.0",
     "shortid": "^2.2.14",
     "styled-components": "^4.2.1"
   },
