+++ src/Containers/Main/Main.js
@@ -1,17 +1,15 @@
 import React, { Component } from 'react';
-import { cleanBooks } from '../../Utility/cleanBooks';
-const API_KEY =`${process.env.REACT_APP_NEWYORKTIMES_API_KEY}`
+// import { cleanBooks } from '../../Utility/cleanBooks';
+
 
 
 export default class Main extends Component {
 
   async componentDidMount(){
-    const url = 
-    fetchAllBooks()
-    // const response = await fetch(`)
-    const books = await response.json()
-    const cleanedBooks = cleanBooks(books)
-    console.log(cleanedBooks)
+    // fetchAllBooks()
+
+  
+    // console.log(cleanedBooks)
   }
 
   render() {
