+++ src/index.js
@@ -5,11 +5,14 @@ import App from './Components/App/App';
 import * as serviceWorker from './serviceWorker';
 import { BrowserRouter } from 'react-router-dom'
 import { Provider } from 'react-redux'
-import { createStore } from 'redux';
+import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension'
 import  rootReducer  from './reducers/index';
+import thunk from 'redux-thunk'
 
-const store = createStore(rootReducer, composeWithDevTools())
+const store = createStore(rootReducer, 
+  composeWithDevTools(applyMiddleware(thunk))
+  )
 
 
 const router = (
